% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accurateData-method.R
\name{accurateData}
\alias{accurateData}
\alias{accurateData,Binalysis-method}
\title{Extract accurate mass data}
\usage{
accurateData(x)

\S4method{accurateData}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\description{
Extract accurate data mass from a Binalysis object.
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
