% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMasses.R
\name{addMasses}
\alias{addMasses}
\title{Add in masses with zero intensity if not present in a peak matrix in a peak list}
\usage{
addMasses(x)
}
\arguments{
\item{x}{A list containing peak matrices for each scan}
}
\value{
A list containing peak matrices with added 0 values
}
\description{
This function will add rows to a peak matrix of missing m/z identified within the supplied peak list to ensure equal matrix row lengths for further processing
}
\author{
Jasen Finch
}

