% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binnedData-method.R
\docType{methods}
\name{binnedData}
\alias{binnedData}
\alias{binnedData,Binalysis-method}
\title{binnedData-Binalysis}
\usage{
binnedData(x)

\S4method{binnedData}{Binalysis}(x)
}
\arguments{
\item{x}{Binalysis object}
}
\description{
Extract binned data from a Binalysis object.
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
