% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{binneR-package}
\alias{binneR}
\alias{binneR-package}
\title{binneR: Spectral Processing for High Resolution Flow Infusion Mass Spectrometry}
\description{
A spectral binning approach for flow infusion electrospray high resolution mass spectrometry (FIE-HRMS) metabolome fingerprinting data. The methodology for this approach is outlined in Finch et al. (2022) \doi{10.1007/s11306-022-01923-6}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://aberhrml.github.io/binneR/}
  \item Report bugs at \url{https://github.com/aberHRML/binneR/issues}
}

}
\author{
\strong{Maintainer}: Jasen Finch \email{jsf9@aber.ac.uk}

}
