% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{binner}
\alias{binner}
\alias{binner-package}
\title{Import and process ABI data files for AFLP analyses}
\description{
Import and process ABI data files for AFLP analyses
}
\details{
\tabular{ll}{
Package: \tab binner\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-10-09\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

Provides functions for reading fsa files
}
\author{
Tyler Smith \email{tyler@plantarum.ca}
}
\keyword{abi}
\keyword{aflp}
\keyword{fsa}
\keyword{package}

