% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{BinParameters-class}
\alias{BinParameters-class}
\title{Spectral binning parameters class}
\description{
An S4 class to store spectral binning parameters.
}
\section{Slots}{

\describe{
\item{\code{scans}}{numeric vector containing the scan indexes to use for binning}

\item{\code{cls}}{the column of class labels to use for aggregating accurate
mass data. Defaults to NULL where accurate mass data will be averaged
across all samples}
}}

\seealso{
\code{\link{binParameters}}
}
