% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binParameters.R
\name{binParameters}
\alias{binParameters}
\title{Set spectral binning parameters}
\usage{
binParameters(scans = 5:12, cls = character())
}
\arguments{
\item{scans}{numeric vector containing the scan indexes to use for binning}

\item{cls}{the column of class labels to use for aggregating accurate
mass data. Defaults to NULL where accurate mass data will be averaged
accross all samples}
}
\value{
S4 object of class BinParameters
}
\description{
Selection of parameters to use for spectral binning.
}
\examples{
p <- binParameters(scans = 6:17)
}
\seealso{
\code{\link{BinParameters-class}}, \code{\link{scans}},
\code{\link{cls}}
}
