% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\docType{methods}
\name{cls}
\alias{cls}
\alias{cls,BinParameters-method}
\alias{cls<-}
\title{cls}
\usage{
cls(x)

\S4method{cls}{BinParameters}(x)

cls(x) <- value
}
\arguments{
\item{x}{an S4 object of class BinParameters}

\item{value}{character containing the column of class labels to use for aggregating accurate mass data}
}
\description{
get and set class labels of a BinParameters object.
}
