% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\docType{methods}
\name{clusterType}
\alias{clusterType}
\alias{clusterType,BinParameters-method}
\title{clusterType}
\usage{
clusterType(x)

\S4method{clusterType}{BinParameters}(x)
}
\arguments{
\item{x}{an S4 object of class BinParameters}
}
\description{
get and set cluster type of a BinParameters object.
}
