% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colMasses.R
\name{colMasses}
\alias{colMasses}
\title{Collate all masses present in the supplied list}
\usage{
colMasses(x)
}
\arguments{
\item{x}{A peak list containing peak matrices for each scan}
}
\value{
A vector containing all the masses present in the supplied peak list
}
\description{
Collate all masses present across all scans in the supplied peak list, into a single reference vector.
}
\author{
Jasen Finch
}

