% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combScans.R
\name{combScans}
\alias{combScans}
\title{Combine scan ranges and retrieve given scans}
\usage{
combScans(x, scans, sranges, modes)
}
\arguments{
\item{x}{A peak list containing peak matrices for each scan}

\item{scans}{A vector of scan numbers that should be retrieved}

\item{sranges}{A list of vectors containing the scan events present}

\item{modes}{A vector of strings denoting mode names}
}
\value{
A list containing peak lists for the relevant scans with combined scan ranges for each present mode.
}
\description{
A function that will split a peak list into appropriate modes, retrieve peak matrices of the supplied relevant scans and combine scan ranges.
}
\author{
Jasen Finch
}

