% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectClusterType}
\alias{detectClusterType}
\title{Detect suitable cluster type}
\usage{
detectClusterType()
}
\description{
Detect appropriate cluster type from OS for parallel processing.
}
\examples{
detectClusterType()
}
