% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectParameters}
\alias{detectParameters}
\title{Detect suitable spectral binning parameters}
\usage{
detectParameters(files)
}
\arguments{
\item{files}{character vector of file paths}
}
\value{
S4 object of class BinParameters
}
\description{
Detect binning parameters from a given list of file paths.
}
\examples{
file_paths <- system.file("example-data/1.mzML.gz", package = "binneR")
parameters <- detectParameters(file_paths)
}
\seealso{
\code{\link{BinParameters-class}}, \code{\link{binParameters}}
}
