% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectSranges}
\alias{detectSranges}
\title{detectSranges}
\usage{
detectSranges(files, nCores = detectCores() * 0.75,
  clusterType = detectClusterType())
}
\arguments{
\item{files}{character vector of file paths.}

\item{nCores}{the number of cores to use for parallel processing}

\item{clusterType}{the type of cluster to use for parallel processing}
}
\description{
Detect scan ranges from file list.
}
