% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fsaDrop}
\alias{fsa2PeakTab}
\alias{fsaCombine}
\alias{fsaDrop}
\alias{fsaNormalize}
\alias{fsaRename}
\alias{plot.electropherogram}
\alias{plot.fsa}
\alias{print.fsa}
\alias{summary.fsa}
\title{View and manipulate fsa files}
\usage{
fsaDrop(fsa, epn)

fsaRename(fsa, epn, newname)

fsaCombine(fsa1, fsa2)

fsaNormalize(fsa)

fsa2PeakTab(fsa, dye)

\method{print}{fsa}(fsa)

\method{summary}{fsa}(fsa)

\method{plot}{fsa}(fsa, epn, ...)

\method{plot}{electropherogram}(ep, raw = FALSE, ...)
}
\arguments{
\item{fsa}{An fsa object, as produced by \code{\link{readFSA}}.}

\item{epn}{The index of one or more samples. Can be an integer or a
character string. Integers refer to samples in the order they are
stored in an \code{fsa} object; character strings refer to them by
name.}

\item{newname}{The new name that will be applied to sample \code{epn}.}

\item{fsa1}{An fsa object, as produced by \code{readFSA}}

\item{fsa2}{An fsa object, as produced by \code{readFSA}}

\item{dye}{Which dye to read when converting to a Peakscanner table.
Valid values include "d6.FAM", "VIC, "NED" and "PET".}
}
\value{
\code{fsaDrop} returns an \code{fsa} object with the indicated samples
removed.

\code{fsaRename} returns an \code{fsa} object with the indicated sample
renamed.

\code{fsaNormalize} returns an \code{fsa} with the rfu values
normalized. The values are scaled such that the sum of the rfus for
each sample will equal the mean of the sum for all samples. Different
dyes are normalized separately.

\code{plot.fsa} is the most convenient way to display an
electropherogram. The \code{epn} argument allows you to pick which
sample you wish to plot, and the \code{raw} argument determines whether
you scale the lines by the raw reads (\code{raw = TRUE}), or by `base
pairs' (\code{raw = FALSE}, the default).

\code{fsa2PeakTab} converts the \code{fsa} to a peak table
\code{data.frame} in a layout similar to that produced by peakscanner.
The data.frame has three columns: sample.name, bp, and rfu. Each row
includes the data for a single peak detected for a single sample. The
rest of the chromatogram data is dropped.
}
\description{
Functions to view electropherograms, rename or remove
samples, combine fsa objects and normalize data.
}
\details{
\code{fsaDrop} removes a sample from an \code{fsa} object.

\code{fsaRename} renames a sample in an \code{fsa} object.

\code{fsaCombine} combines two \code{fsa} objects

\code{fsaNormalize} normalizes the rfu values across each sample in
the fsa.

\code{fsa2PeakTab} converts the fsa to a peakscanner-style peaks
table.
}
\examples{
\dontrun{
## The following examples assume fsa.data is an fsa object created by a
## call to readFSA()

## Remove the tenth sample
fsa.data.B <- fsaDrop(fsa.data, 10)

## Remove sample `unit3'
fsa.data.C <- fsaDrop(fsa.data, "unit3")

## Rename sample `d.x11' to `unit4'
fsa.data.D <- fsaRename(fsa.data, epn = "d.x11", newname = "unit4")

## Combine fsa objects from two different file sets
fsa.combined <- fsaCombine(fsa1 = block1, fsa2 = block2)

## Normalize the RFU data across all samples in an fsa object
fsa.norm <- fsaNormalize(fsa.data)

}
}
\author{
Tyler Smith
}
\keyword{aflp}
\keyword{fsa}
\keyword{genemapper}
\keyword{microsatellite}
\keyword{peakscanner}
\keyword{ssr}

