% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/info-method.R
\name{info}
\alias{info}
\alias{info,Binalysis-method}
\title{Extract sample information}
\usage{
info(x)

\S4method{info}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\description{
Extract runinfo data from a Binalysis object.
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
