% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massMat.R
\name{massMat}
\alias{massMat}
\title{build an intensity matrix for supplied peak list}
\usage{
massMat(x)
}
\arguments{
\item{x}{A peak list containing peak matrices for each scan of equal row length}
}
\value{
An intensity matrix with as many rows as peak matrices in peak list
}
\description{
build and intensity matrix for a supplied peak list. Can be used to build intensity matrices for combining either scans within a sample of multiple samples.
}
\author{
Jasen Finch
}

