% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\docType{methods}
\name{modes}
\alias{modes}
\alias{modes,BinParameters-method}
\alias{modes<-}
\title{modes}
\usage{
modes(x)

\S4method{modes}{BinParameters}(x)

modes(x) <- value
}
\arguments{
\item{x}{an S4 object of class BinParameters}

\item{value}{character vector of mode indexes}
}
\description{
get and set ionisation modes of a BinParameters object.
}
