% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{oxyPT}
\alias{oxyPT}
\title{The peak table for 203 wild cranberry samples from Ontario and Quebec.}
\format{A data frame with 13364 rows and 3 variables:
\describe{
  \item{sample.name}{The individual plant sample}
  \item{bp}{Peak size in base pairs}
  \item{height}{The peak height in (normalized) RFUs}
}}
\source{
Smith et al., In Review. Evaluating the relationship between
diploid and tetraploid Vaccinium oxycoccos L. Ericaceae in eastern Canada.
}
\usage{
oxyPT
}
\description{
Cranberry AFLP data from Smith et al. In Press. Generated using binner
from raw fsa files.
}
\keyword{datasets}

