% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\name{scans}
\alias{scans}
\alias{scans<-}
\alias{cls}
\alias{cls<-}
\alias{nCores}
\alias{nCores<-}
\alias{clusterType}
\alias{clusterType<-}
\alias{scans,BinParameters-method}
\alias{scans<-,BinParameters-method}
\alias{cls,BinParameters-method}
\alias{cls<-,BinParameters-method}
\alias{nCores,BinParameters-method}
\alias{nCores<-,BinParameters-method}
\alias{clusterType,BinParameters-method}
\alias{clusterType<-,BinParameters-method}
\title{Get and set spectral binning parameters}
\usage{
scans(x)

scans(x) <- value

cls(x)

cls(x) <- value

nCores(x)

nCores(x) <- value

clusterType(x)

clusterType(x) <- value

\S4method{scans}{BinParameters}(x)

\S4method{scans}{BinParameters}(x) <- value

\S4method{cls}{BinParameters}(x)

\S4method{cls}{BinParameters}(x) <- value

\S4method{nCores}{BinParameters}(x)

\S4method{nCores}{BinParameters}(x) <- value

\S4method{clusterType}{BinParameters}(x)

\S4method{clusterType}{BinParameters}(x) <- value
}
\arguments{
\item{x}{S4 object of class BinParameters}

\item{value}{parameter value to set}
}
\description{
Get and set spectral binning parameters.
}
