% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotChromFromFile}
\alias{plotChromFromFile}
\title{Plot an infusion profile chromatogram from a file}
\usage{
plotChromFromFile(files, scans = c())
}
\arguments{
\item{files}{character vector of file paths to use}

\item{scans}{specify scans to highlight within the plot}
}
\description{
Plot and averaged infusion profile from a vector of specified
file paths.
}
\examples{
file_paths <- system.file("example-data/1.mzML.gz", package = "binneR")

plotChromFromFile(file_paths,
  scans = detectInfusionScans(file_paths)
)
}
