% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotFingerprint.R
\name{plotFingerprint}
\alias{plotFingerprint}
\alias{plotFingerprint,Binalysis-method}
\title{plotFingerprint}
\usage{
plotFingerprint(x)

\S4method{plotFingerprint}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\description{
Plot averaged spectrum fingerprint.
}
