% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotTIC.R
\docType{methods}
\name{plotTIC}
\alias{plotTIC}
\alias{plotTIC,Binalysis-method}
\title{plotTIC}
\usage{
plotTIC(x, by = "injOrder", colour = "block")

\S4method{plotTIC}{Binalysis}(x, by = "injOrder", colour = "block")
}
\arguments{
\item{x}{S4 object of class Binalysis}

\item{by}{info column to plot against}

\item{colour}{info column to provide colour labels}
}
\description{
Plot sample total ion counts.
}
