% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fsaRGbin}
\alias{fsaRGbin}
\title{Locate bins using the RawGeno algorithm}
\usage{
fsaRGbin(pt, start = 49.999, end = 495, mxbin = 1.5, mnbin = 1,
  verbose = FALSE)
}
\arguments{
\item{pt}{a peak table, as produced by \code{fsa2PeakTab}}
}
\value{
TBA
}
\description{
\code{fsaRGbin} identifies bins in a peak table
}
\details{
fsaRGbin calculates bins using all peaks in a peak table, including
very small ones. As a consequence, some bins may include only peaks <
50 rfus, and will therefore appear to be 'empty' in \code{scanGel()}.
These empty bins should be filtered out before analysis, and are not
otherwise of any concern.

Note that the way the algorithm is coded, the largest bin will always
be bounded on the upper side by 495. Consequently, this bin may appear
quite large. It will only ever include fragments that are collectively
within mxbin base pairs of each other, but may include a sizeable empty
section above the actual fragments. Again, this is not a concern for
subsequent analysis.
}
\examples{
TBA
}
\author{
Tyler Smith
}
\references{
This function uses the `RawGeno' binning algorithm of
\href{http://www.biomedcentral.com/1471-2105/10/33}{Arrigo et al.
2009}. I reimplemented this following their flow-chart, rather than
directly copying the code in the RawGeno package. As a consequence,
there may be some small numerical differences between the two
implementations. I elected not to copy the code as the style of the two
packages are not easily reconciled.

Arrigo, N., Tuszynski, J. W., Ehrich, D., Gerdes, T., & Alvarez, N.
(2009). Evaluating the impact of scoring parameters on the structure of
intra-specific genetic variation using RawGeno, an R package for
automating AFLP scoring. BMC bioinformatics, 10(1), 33.
}
\keyword{aflp}
\keyword{binning}

