% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{version}
\alias{version}
\alias{creationDate}
\alias{filePaths}
\alias{sampleInfo}
\alias{binnedData}
\alias{accurateData}
\alias{binningParameters}
\alias{version,Binalysis-method}
\alias{creationDate,Binalysis-method}
\alias{filePaths,Binalysis-method}
\alias{sampleInfo,Binalysis-method}
\alias{binnedData,Binalysis-method}
\alias{accurateData,Binalysis-method}
\alias{binningParameters,Binalysis-method}
\title{Binalysis class get methods}
\usage{
version(x)

creationDate(x)

filePaths(x)

sampleInfo(x)

binnedData(x)

accurateData(x)

binningParameters(x)

\S4method{version}{Binalysis}(x)

\S4method{creationDate}{Binalysis}(x)

\S4method{filePaths}{Binalysis}(x)

\S4method{sampleInfo}{Binalysis}(x)

\S4method{binnedData}{Binalysis}(x)

\S4method{accurateData}{Binalysis}(x)

\S4method{binningParameters}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}

\item{value}{value to set}
}
\description{
Methods for accessing spectral binning results from the  Binalysis S4 class.
}
\details{
\itemize{
 \item{version}{ - Extract package version used to create the object.}
 \item{creationDate}{ - Extract the creation date of the object.}
 \item{filePaths}{ - Extract the sample file paths.}
 \item{sampleInfo}{ - Extract sample meta information.}
 \item{binnedData}{ - Extract a binned data intensity matrix.}
 \item{accurateData}{ - Extract sample-wise accurate mass data and bin measures.}
 \item{binningParametesr}{Extract the binning parameters.}
}
}
\seealso{
\code{\link{Binalysis-class}}, \code{\link{binneRlyse}}
}
