% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampProcess.R
\name{sampProcess}
\alias{sampProcess}
\title{process a single data file}
\usage{
sampProcess(file, scans, dp, sranges, modes)
}
\arguments{
\item{file}{File path of converted data file}

\item{scans}{A vector of scan numbers that should be retrieved}

\item{dp}{An integer denoting the number of decimal places for spectral binning}

\item{sranges}{A list of vectors containing the scan events present}

\item{modes}{A vector of strings denoting mode names}
}
\value{
A list containing peak lists for the relevant scans with combined scan ranges for each present mode in the data file.
}
\description{
Apply spectral binning on a single data file.
}
\author{
Jasen Finch
}

