% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scanFSA}
\alias{fsa.plot}
\alias{scanFSA}
\alias{scroll.fsa}
\title{Visualize fsa data}
\usage{
scanFSA(fsa, sample, bp = c(50, 75), height = c(0, 3000), chan = 1)

scroll.fsa(fsa, sample, bp = c(35, 500), height = c(0, 3000), chan = 1)

fsa.plot(fsa, sample, chan = 1, raw = FALSE, bp = c(50, 500),
  ylim = NULL, scans = c(0, 8000), ...)
}
\arguments{
\item{fsa}{an fsa file, as produced by \code{read.fsa}}

\item{sample}{The sample to plot. If numeric, it will plot the samples with the
corresponding levels for the tag column. i.e., \code{samples = 3:5} will plot samples
corresponding to levels(fsa$dat$tag)[3:5].

If \code{sample} is a single string, all samples that contain that string will be
plotted.

Otherwise, \code{sample} should be a list of the tags to display.}

\item{bp}{The range in base pairs to display (i.e., the x-axis limits)}

\item{height}{The range of rfus to display (i.e., the y-axis limits)}

\item{chan}{The data channel to display. '1' is the blue/FAM channel.}

\item{raw}{If TRUE, display the fragment size as read times, rather than base pairs.
With this option on, the ladder for the first sample is overlayed on the plot. Note
that if you plot more than one sample this way, the fragment sizes will not be
standardized among them. This option is best used with a single sample.}
}
\value{
Nothing of interest. Only used for plotting.
}
\description{
\code{scanFSA} provides an interactive plot of fsa chromatograms.
\code{fsa.plot} provides a single static plot of fsa data.
}
\author{
Tyler Smith
}
\keyword{aflp}
\keyword{fsa}
\keyword{plot}

