% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\name{scans}
\alias{scans}
\alias{scans,BinParameters-method}
\alias{scans<-}
\title{scans}
\usage{
scans(x)

\S4method{scans}{BinParameters}(x)

scans(x) <- value
}
\arguments{
\item{x}{an S4 object of class BinParameters}

\item{value}{numeric vector of scan numbers}
}
\description{
get and set scans of a BinParameters object.
}
