% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,Binalysis-method}
\alias{show,Binalysis-method}
\title{show-Binalysis}
\usage{
\S4method{show}{Binalysis}(object)
}
\arguments{
\item{object}{Binalysis Object}
}
\description{
show method for Binalysis class
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
