% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show,BinParameters-method}
\alias{show,BinParameters-method}
\title{show-BinParameters}
\usage{
\S4method{show}{BinParameters}(object)
}
\arguments{
\item{object}{BinParameters Object}
}
\description{
show method for BinParameters class
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
