% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\name{show,BinParameters-method}
\alias{show,BinParameters-method}
\alias{show,Binalysis-method}
\title{Show methods for spectral binning classes}
\usage{
\S4method{show}{BinParameters}(object)

\S4method{show}{Binalysis}(object)
}
\arguments{
\item{object}{S4 objects of class BinParameters or Binanalysi}
}
\description{
Show methods for spectral binning classes.
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
