% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\docType{methods}
\name{sranges}
\alias{sranges}
\alias{sranges,BinParameters-method}
\alias{sranges<-}
\title{sranges}
\usage{
sranges(x)

\S4method{sranges}{BinParameters}(x)

sranges(x) <- value
}
\arguments{
\item{x}{an S4 object of class BinParameters}

\item{value}{list containing low and high scan range vectors}
}
\description{
get and set scan ranges of a BinParameters object.
}
