% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewSpectrum.R
\name{viewSpectrum}
\alias{viewSpectrum}
\title{Shiny app to view}
\usage{
viewSpectrum(file = NULL, polarityOrder = NULL)
}
\arguments{
\item{file}{File path of data file to open}

\item{polarityOrder}{specified only when using mzML files. A numeric vector of the length of the number of polarities present in the file giving the polarity order.}
}
\description{
A shiny app to view converted MS data files.
}
\examples{
\dontrun{
# for an mzXML file
viewSpectrum(file=list.files(system.file('mzXML',package='binneR'),full.names = T)[1])

# for an mzML file containing both polarities with negative scanning before positive
viewSpectrum(polarityOrder=c(0,1))
}
}
\author{
Jasen Finch
}

