% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accurateData-method.R
\name{accurateData}
\alias{accurateData}
\alias{accurateData,Binalysis-method}
\title{Extract accurate mass data}
\usage{
accurateData(x)

\S4method{accurateData}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\value{
A tibble containing sample wise accurate mass data.
}
\description{
Extract sample-wise accurate mass data and bin measures for each 0.01 amu bin from a Binalysis object.
}
\seealso{
\code{\link{binneRlyse}}
}
