% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/accurateData-method.R
\docType{methods}
\name{accurateData}
\alias{accurateData}
\alias{accurateData,Binalysis-method}
\title{accurateData-Binalysis}
\usage{
accurateData(x)

\S4method{accurateData}{Binalysis}(x)
}
\arguments{
\item{x}{Binalysis object}
}
\description{
Extract accurate data from a Binalysis object.
}
\author{
Jasen Finch \email{jsf9@aber.ac.uk}
}
