% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binneRlyse.R
\name{binneRlyse}
\alias{binneRlyse}
\title{Perform spectral binning}
\usage{
binneRlyse(files, info, parameters = binParameters(), verbose = TRUE)
}
\arguments{
\item{files}{character vector of file paths to use for spectral binning}

\item{info}{tibble containing sample information}

\item{parameters}{object of class BinParameters containing parameters 
for spectral binning}

\item{verbose}{show console output}
}
\value{
S4 object of class Binalysis.
}
\description{
perform spectral binning.
}
\examples{
\dontrun{
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')

info <- metaboData::runinfo('FIE-HRMS','BdistachyonEcotypes')

parameters <- detectParameters(files)
cls(parameters) <- 'class'

analysis <- binneRlyse(files, 
                       info, 
                       parameters = parameters)
   }
}
\seealso{
\code{\link{Binalysis-class}}, \code{\link{binParameters}}, \code{\link{info}}, \code{\link{binnedData}},  \code{\link{accurateData}}
}
