% Generated by roxygen2 (4.0.2): do not edit by hand
\name{binSet}
\alias{aflpCombine}
\alias{binSet}
\title{Apply bins to peak tables, creating arrays of AFLP data}
\usage{
binSet(pt, bins, pref, hthresh = 100)

aflpCombine(aflp1, aflp2)
}
\arguments{
\item{pt}{A peak table, as generated by \code{fsa2PeakTab}}

\item{bins}{A bin table, as produced by \code{fsaRGbin}}

\item{pref}{A prefix to append to bin names, to allow for later
combining data from different primer pairs}

\item{hthresh}{The height threshold to score peaks as 'present'. You
can revisit this decision later in the process.}

\item{aflp1}{aflp data arrays, as produced by \code{binSet}}

\item{aflp2}{aflp data arrays, as produced by \code{binSet}}
}
\value{
\code{binSet} returns a 3D array. The rows are
samples, the columns are fragments, and the third dimension
splits the array into allele calls ("alleles"), peak heights
("heights") and fragment size in basepairs ("sizes"). Thus,
to retrieve the presence/absence matrix of allele calls, use:
\code{aflp[ , , "alleles"]}, where \code{aflp} is the object
returned by \code{binSet}.

Other data summaries are readily available by creative slicing and
dicing of the 3D array.
}
\description{
Functions for converting peak tables to AFLP data arrays,
and combining multiple arrays for analysis.
}
\details{
\code{binSet} takes a peak table, applies the
given bins, and returns a 3D aflp data array.

\code{aflpCombine} takes two aflp data arrays, as produced from
\code{binSet}, and combines them into a single array
}
\author{
Tyler Smith
}

