% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectInfusionScans}
\alias{detectInfusionScans}
\title{Detect infusion profile scan range}
\usage{
detectInfusionScans(files, thresh = 0.5)
}
\arguments{
\item{files}{character vector of file paths to use}

\item{thresh}{detection threshold as a proportion of the peak of the 
infusion profile}
}
\description{
Detect infusion scans for a set of FIE-MS infusion profiles.
}
\examples{
if (requireNamespace("metaboData", quietly = TRUE)) {
   detectInfusionScans(
      metaboData::filePaths('FIE-HRMS',
                            'BdistachyonEcotypes')[1])
}
}
