% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectInfusionScans}
\alias{detectInfusionScans}
\title{detectInfusionScans}
\usage{
detectInfusionScans(files, thresh = 0.5, nCores = detectCores() * 0.75,
  clusterType = detectClusterType())
}
\arguments{
\item{files}{character vector of file paths to use}

\item{thresh}{detection threshold as a proportion of  preak of the infusion profile}

\item{nCores}{the number of cores to use for parallel processing}

\item{clusterType}{the type of cluster to use for parallel processing}

\item{sranges}{A list of vectors containing the scan events present.}
}
\description{
detect infusion scans for a set of FIE-MS infusion profiles.
}
\examples{
if (requireNamespace("metaboData", quietly = TRUE)) {
   detectInfusionScans(metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1])
}
}
