% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectInfusionScans.R
\name{detectInfusionScans}
\alias{detectInfusionScans}
\title{detectInfusionScans}
\usage{
detectInfusionScans(files, sranges = list(c(70, 1000)), thresh = 0.5)
}
\arguments{
\item{files}{character vector of file paths to use}

\item{sranges}{A list of vectors containing the scan events present.}

\item{thresh}{detection threshold as a proportion of  preak of the infusion profile}
}
\description{
detect infusion scans for a set of FIE-MS infusion profiles.
}
\examples{
if (requireNamespace("metaboData", quietly = TRUE)) {
   detectInfusionScans(metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1])
}
}
