% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{detectParameters}
\alias{detectParameters}
\title{Detect suitable spectral binning parameters}
\usage{
detectParameters(files)
}
\arguments{
\item{files}{character vector of file paths}
}
\description{
Detect binning parameters from a given list of file paths.
}
\examples{
files <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')
parameters <- detectParameters(files[1])
}
