% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/binParameters.R
\docType{methods}
\name{nCores}
\alias{nCores}
\alias{nCores,BinParameters-method}
\alias{nCores<-}
\title{nCores}
\usage{
nCores(x)

\S4method{nCores}{BinParameters}(x)

nCores(x) <- value
}
\arguments{
\item{x}{an S4 object of class BinParameters}

\item{value}{number of cores to use for parallel processing}
}
\description{
get and set number of cores of a BinParameters object.
}
