% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plot.R
\name{plotBin}
\alias{plotBin}
\alias{plotBin,Binalysis-method}
\title{Plot a spectral bin feature}
\usage{
plotBin(x, bin, type = c("all", "cls", "sample"))

\S4method{plotBin}{Binalysis}(x, bin, type = c("all", "cls", "sample"))
}
\arguments{
\item{x}{S4 object of class Binalysis}

\item{bin}{0.01amu bin to plot}

\item{type}{bin to be plotted as a total (all), class (cls) or sample spectra.}
}
\description{
Kernal density plot of a specified spectral bin feature.
}
\seealso{
\code{\link{accurateData}}, \code{\link{binneRlyse}}
}
