% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotBin.R
\docType{methods}
\name{plotBin}
\alias{plotBin}
\alias{plotBin,Binalysis-method}
\title{plotBin}
\usage{
plotBin(x, bin, cls = T)

\S4method{plotBin}{Binalysis}(x, bin, cls = T)
}
\arguments{
\item{x}{S4 object of class Binalysis}

\item{bin}{0.01amu bin to plot}

\item{cls}{\code{TRUE} or \code{FALSE}. Should bins be plotted by class?}
}
\description{
kernal density plot of a specified bin.
}
