% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotMeasures.R
\name{plotCentrality}
\alias{plotCentrality}
\alias{plotCentrality,Binalysis-method}
\title{plotCentrality}
\usage{
plotCentrality(x, histBins = 30)

\S4method{plotCentrality}{Binalysis}(x, histBins = 30)
}
\arguments{
\item{x}{S4 object of class Binalysis}

\item{histBins}{number of bins to use for histogram plotting}
}
\description{
Plot the bin centrality distribution for a Binalysis object.
}
