% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plot.R
\name{plotChromatogram}
\alias{plotChromatogram}
\alias{plotChromatogram,Binalysis-method}
\title{Plot an infusion profile chromatogram}
\usage{
plotChromatogram(x)

\S4method{plotChromatogram}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class \code{Binalysis}}
}
\description{
Plot an averaged infusion profile chromatogram from a Binalysis object.
}
\seealso{
\code{\link{binneRlyse}}
}
