% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotChromatogram.R
\docType{methods}
\name{plotChromatogram}
\alias{plotChromatogram}
\alias{plotChromatogram,Binalysis-method}
\title{plotChromatogram}
\usage{
plotChromatogram(x)

\S4method{plotChromatogram}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class \code{Binalysis}}
}
\description{
plot an averaged infusion profile from a Binalysis object.
}
