% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotChromFromFile}
\alias{plotChromFromFile}
\title{Plot an infusion profile chromatogram from a file}
\usage{
plotChromFromFile(files, scans = c())
}
\arguments{
\item{files}{character vector of file paths to use}

\item{scans}{specify scans to highlight within the plot}
}
\description{
Plot and averaged infusion profile from a vector of specified 
file paths.
}
\examples{
plotChromFromFile(metaboData::filePaths('FIE-HRMS','BdistachyonTechnical')[1],
                            scans = c(6,18))
}
