% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChromatogram.R
\name{plotChromFromFile}
\alias{plotChromFromFile}
\title{plotChromFromFile}
\usage{
plotChromFromFile(files, sranges = list(c(70, 1000)), scans = c())
}
\arguments{
\item{files}{character vector of file paths to use}

\item{sranges}{A list of vectors containing the scan events present}

\item{scans}{specify scans to highlight within the plot}
}
\description{
plot and averaged infusion profile from a vector of specified file paths.
}
\examples{
plotChromFromFile(list.files(system.file('mzML',package = 'binneR'),
                            full.names=TRUE),scans = c(6,18))
}
