% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/plotMeasures.R
\name{plotPurity}
\alias{plotPurity}
\alias{plotPurity,Binalysis-method}
\title{plotPurity}
\usage{
plotPurity(x, histBins = 30)

\S4method{plotPurity}{Binalysis}(x, histBins = 30)
}
\arguments{
\item{x}{S4 object of class Binalysis}
}
\description{
Plot the bin purity distribution for a Binalysis object.
}
