% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fsaRGbin}
\alias{fsaRGbin}
\title{Locate bins using the RawGeno algorithm}
\usage{
fsaRGbin(pt, start = 49.999, end = 495, mxbin = 1.5, mnbin = 1,
  verbose = FALSE)
}
\arguments{
\item{pt}{a peak table, as produced by \code{fsa2PeakTab}}
}
\value{
TBA
}
\description{
\code{fsaRGbin} identifies bins in a peak table
}
\details{
TBA
}
\examples{
TBA
}
\author{
Tyler Smith
}
\references{
This function uses the `RawGeno' binning algorithm of
\href{http://www.biomedcentral.com/1471-2105/10/33}{Arrigo et al.
2009}. I reimplemented this following their flow-chart, rather than
directly copying the code in the RawGeno package. As a consequence,
there may be some small numerical differences between the two
implementations. I elected not to copy the code as the style of the two
packages are not easily reconciled.

Arrigo, N., Tuszynski, J. W., Ehrich, D., Gerdes, T., & Alvarez, N.
(2009). Evaluating the impact of scoring parameters on the structure of
intra-specific genetic variation using RawGeno, an R package for
automating AFLP scoring. BMC bioinformatics, 10(1), 33.
}
\keyword{aflp}
\keyword{binning}

