% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBinningParameters.R
\name{readBinningParameters}
\alias{readBinningParameters}
\title{Read binning parameters from file}
\usage{
readBinningParameters(file)
}
\arguments{
\item{file}{file path}
}
\value{
S4 object of class BinParameters.
}
\description{
Parse binning parameters from a YAML formatted file.
}
\examples{
file <- system.file('binning_parameters.yml',package = 'binneR')
parameters <- readBinningParameters(file)
}
\seealso{
\code{\link{BinParameters-class}}
}
