% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Read and process mulitple data files}
\usage{
readFiles(files, dp, scans, sranges = list(c(50, 1000)), modes = c("p",
  "n"), nCores = 2)
}
\description{
Read and process mulitple data files
}

