% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFiles.R
\name{readFiles}
\alias{readFiles}
\title{Read and process mulitple data files}
\usage{
readFiles(files, dp, scans)
}
\arguments{
\item{files}{A vector of converted data file paths}

\item{dp}{An integer denoting the number of decimal places for spectral 
binning}

\item{scans}{A vector of scan numbers that should be retrieved}

\item{nCores}{The number of cores on which to parallel process.}

\item{clusterType}{the type of cluster to use for parallel processing}
}
\value{
A list containing peak lists for the relevant scans with combined 
scan ranges for each present mode in the data file.
}
\description{
Apply spectral binning on multiple data files.
}
\details{
Parallel processing is managed by the \code{future} package. This can be specified using the \code{plan() function}. See the example below and \code{?future::plan} for details on how this can be specified.
}
\examples{
## Example file path
file_paths <- metaboData::filePaths('FIE-HRMS',
                                    'BdistachyonEcotypes')[1]

## Optionally use declare parallel processing options
plan(future::multisession,workers = 2)
                                                                        
## Process example file
res <- readFiles(file_paths,
                 dp = 2,
                 scans = 6:17)
}
\author{
Jasen Finch
}
