% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/access.R
\name{sampleInfo}
\alias{sampleInfo}
\alias{binnedData}
\alias{accurateData}
\alias{sampleInfo,Binalysis-method}
\alias{binnedData,Binalysis-method}
\alias{accurateData,Binalysis-method}
\title{Binalysis class get methods}
\usage{
sampleInfo(x)

binnedData(x)

accurateData(x)

\S4method{sampleInfo}{Binalysis}(x)

\S4method{binnedData}{Binalysis}(x)

\S4method{accurateData}{Binalysis}(x)
}
\arguments{
\item{x}{S4 object of class Binalysis}

\item{value}{value to set}
}
\description{
Methods for accessing spectral binning results from the  Binalysis S4 class.
}
\details{
\itemize{
 \item{sampleInfo}{Extract sample meta information.}
 \item{binnedData}{Extract a binned data intensity matrix.}
 \iten{accurateData}{Extract sample-wise accurate mass data and bin measures.}
}
}
\seealso{
\code{\link{Binalysis-class}}, \code{\link{binneRlyse}}
}
