% Generated by roxygen2 (4.0.2): do not edit by hand
\name{scanGel}
\alias{scanGel}
\title{Gel-like visualization of electropherograms, and tools to
interactively manipulate bin boundaries}
\usage{
scanGel(pt, bin.lines = numeric())
}
\arguments{
\item{pt}{A peak table, as produced by \code{fsa2PeakTab}}

\item{bin.lines}{The bin boundaries to use. Can be a two column table,
with the first column being the lower boundary and the second the upper
boundary; or a vector, with odd elements lower boundaries and even
elements upper boundaries. Omit if you want to define bins completely
manually, otherwise use the output of \code{fsaRGbin}.}
}
\value{
Another placeholder
}
\description{
\code{scanGel} plots the raw data from an fsa file.
}
\details{
This is a placeholder
}
\author{
Tyler Smith
}

