% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSample.R
\name{singleSample}
\alias{singleSample}
\title{singleSample}
\usage{
singleSample(
  file,
  class = NA,
  nCores = detectCores(),
  clusterType = detectClusterType(),
  verbose = TRUE
)
}
\arguments{
\item{file}{file path}

\item{class}{optional class name}

\item{nCores}{number of cores to use for parallel processing}

\item{clusterType}{cluster type to use for parallel processing}

\item{verbose}{show console output}
}
\description{
perform spectral binning on a single sample
}
\examples{
filePath <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1]

bd <- singleSample(filePath)
}
