% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleSample.R
\name{singleSample}
\alias{singleSample}
\title{Perform single sample spectral binning}
\usage{
singleSample(file, class = NA, verbose = TRUE)
}
\arguments{
\item{file}{file path}

\item{class}{optional class name}

\item{verbose}{show console output}
}
\description{
Perform spectral binning on a single sample.
}
\examples{
filePath <- metaboData::filePaths('FIE-HRMS','BdistachyonEcotypes')[1]

bd <- singleSample(filePath)
}
